Provide helpful message if parameter cannot be set.
For private @Parameter fields is users get an exception like
"java.lang.IllegalAccessException: Class ... can not access a member
of class X with modifiers private" The new message "Cannot set
parameter 'parameter'. Ensure that the the field 'parameter' is
public." tells the user what they should do.

The reason for adding this feature is the Stackoverflow question
https://stackoverflow.com/questions/44522046/reflection-exception-in-parameterized-junit-test-using-array-parameter/44522988